requirements.txt
setuptools==49.6.0
Django==3.1.1
djangorestframework==3.11.0
django-rest-framework-social-oauth2==1.1.0
python-dotenv==0.14.0
gunicorn 
whitenoise
dj-database-url
psycopg2